modify log + enable ansi + rename microserviceProduct folder
