+++ src/main/java/com/gftraining/microservice_product/MicroserviceProductApplication.java
@@ -1,9 +1,11 @@
-package com.gftraining.microserviceProduct;
+package com.gftraining.microservice_product;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
+@Slf4j
 public class MicroserviceProductApplication {
 
 	public static void main(String[] args) {
