+++ src/test/java/com/gftraining/microservice_product/MicroserviceProductApplicationTests.java
@@ -1,4 +1,4 @@
-package com.gftraining.microserviceProduct;
+package com.gftraining.microservice_product;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
